depthforge/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "depthforge"
version = "0.1.0"
description = "Professional stereogram engine for VFX, motion graphics, and creative pipelines"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [{ name = "DepthForge Contributors" }]
keywords = ["stereogram", "3d", "vfx", "depth", "sirds", "anaglyph", "comfyui", "nuke", "ofx"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Video",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core deps — the engine runs on these alone
dependencies = [
    "numpy>=1.24",
    "Pillow>=10.0",
]

[project.optional-dependencies]
# Tier 1: Better quality filters + faster processing
cv2 = ["opencv-python>=4.8"]

# Tier 2: Advanced signal processing
scipy = ["scipy>=1.10"]

# Tier 3: AI depth estimation
ai = [
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
]

# Web preview server
web = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "python-multipart>=0.0.6",
]

# Full install — everything
full = [
    "opencv-python>=4.8",
    "scipy>=1.10",
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "python-multipart>=0.0.6",
]

# Dev / testing
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.scripts]
depthforge = "depthforge.cli.main:main"

[project.urls]
Homepage = "https://github.com/depthforge/depthforge"
Documentation = "https://depthforge.readthedocs.io"
Repository = "https://github.com/depthforge/depthforge"
"Bug Tracker" = "https://github.com/depthforge/depthforge/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["depthforge*"]

[tool.setuptools.package-data]
depthforge = [
    "patterns/**/*.png",
    "patterns/**/*.json",
    "presets/*.json",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
ignore_missing_imports = true
